Add devcontainer folder for easy configuration.
By having the devcontainer folder and Dockerfile inside it,
Visual Studio Code automaticaly regocnizes the pattern and
will be prompting us to open the folder inside the container,
for an easier development process.
Optimized the make process for Dockerfile command to use the
number of available processors (nproc command) to speed up
things a bit.